+++ app/src/main/java/one/chandan/rubato/domain/MediaId.kt
@@ -1,0 +1,11 @@
+package one.chandan.rubato.domain
+
+data class MediaId(
+    val sourceType: MediaSourceType,
+    val sourceId: String,
+    val externalId: String
+) {
+    fun asCanonicalId(): String {
+        return "${sourceType.id}:$sourceId:$externalId"
+    }
+}