+++ app/src/main/java/one/chandan/rubato/domain/MediaImage.kt
@@ -1,0 +1,7 @@
+package one.chandan.rubato.domain
+
+data class MediaImage(
+    val id: MediaId? = null,
+    val uri: String? = null,
+    val description: String? = null
+)